LayoutTests/editing/execCommand/insert-paragraph-twice-at-end-of-block-styled-listitem.html

<!DOCTYPE html>
<html>
<head>
<style>
    ul li {display:block; margin-left: 25px;}
</style>
</head>
<body>
    <div id="container" contenteditable="true">
        <p>Paragraph before list.</p>
        <ul>
            <li>List item one</li>
            <li id="target">List item two.</li>
        </ul>
        <p>Paragraph after list.</p>
    </div>
    <script src="../../resources/dump-as-markup.js"></script>
    <script>

        Markup.description('This test verifies that inserting paragraph twice at the end of last listitem styled with display:block/float does not remove entire list.\nExpected behavior is list should not be removed and one empty line should get added.');

        getSelection().collapse(document.getElementById('target'), 1);
        Markup.dump('container', 'Initial');
        document.execCommand("InsertParagraph");
        Markup.dump('container', 'After inserting first paragraph');
        document.execCommand("InsertParagraph");
        Markup.dump('container', 'After inserting second paragraph');
    </script>
</body>
</html>
