Safari-611.1.10/LayoutTests/editing/execCommand/insert-paragraph-twice-at-end-of-block-styled-listitem-expected.txt

This test verifies that inserting paragraph twice at the end of last listitem styled with display:block/float does not remove entire list.
Expected behavior is list should not be removed and one empty line should get added.

Initial:
| "\n        "
| <p>
|   "Paragraph before list."
| "\n        "
| <ul>
|   "\n            "
|   <li>
|     "List item one"
|   "\n            "
|   <li>
|     id="target"
|     "List item two.<#selection-caret>"
|   "\n        "
| "\n        "
| <p>
|   "Paragraph after list."
| "\n    "

After inserting first paragraph:
| "\n        "
| <p>
|   "Paragraph before list."
| "\n        "
| <ul>
|   "\n            "
|   <li>
|     "List item one"
|   "\n            "
|   <li>
|     id="target"
|     "List item two."
|   <li>
|     id="target"
|     <#selection-caret>
|     <br>
|   "\n        "
| "\n        "
| <p>
|   "Paragraph after list."
| "\n    "

After inserting second paragraph:
| "\n        "
| <p>
|   "Paragraph before list."
| "\n        "
| <ul>
|   "\n            "
|   <li>
|     "List item one"
|   "\n            "
|   <li>
|     id="target"
|     "List item two."
|   "\n        "
| <div>
|   <#selection-caret>
|   <br>
| "\n        "
| <p>
|   "Paragraph after list."
| "\n    "
